requirements.txt
picamera==1.13
opencv-python==4.5.1.48
pyyaml==5.4.1
flask
gunicorn
